Initial commit: High-performance cryptocurrency puzzle solver

+++ SECPK1/SECP256K1.cpp
@@ -93,7 +93,7 @@ std::vector<Point> Secp256K1::ComputePublicKeys(std::vector<Int> &privKeys) {
   Int *inv = new Int[privKeys.size()];
   pts.reserve(privKeys.size());
 
-  for(int i=0;i<privKeys.size();i++) {
+  for(int i=0;i<(int)privKeys.size();i++) {
     Point P = ComputePublicKey(&privKeys[i],false);
     inv[i].Set(&P.z);
     pts.push_back(P);
@@ -102,7 +102,7 @@ std::vector<Point> Secp256K1::ComputePublicKeys(std::vector<Int> &privKeys) {
   grp.Set(inv);
   grp.ModInv();
 
-  for(int i = 0; i<privKeys.size(); i++) {
+  for(int i = 0; i<(int)privKeys.size(); i++) {
     pts[i].x.ModMulK1(inv + i);
     pts[i].y.ModMulK1(inv + i);
     pts[i].z.SetInt32(1);
